Day 4: Templates, URLS, slugs; adding images
Learned to setup media folder and add image fields to models, learned how to access image fields in HTML, learned to use path converters to allow one URL to cover many instances of a model being rendered to a template.

learned how named URLS can be accessed from HTML and have parameters passed into them, then into the path() function and finally into the view function that renders a particular template. We can use these parameters for tons of things, including rendering the content of a particular model instance on an HTML page (if you pass it into the context parameter of render() in the view function)